pyproject.toml
[project]
name = "computer-vision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "openai-clip>=1.0.1",
    "opencv-python>=4.11.0.86",
    "pycocotools>=2.0.10",
    "requests>=2.32.5",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "transformers>=4.55.2",
    "ultralytics>=8.3.181",
]

[dependency-groups]
dev = [
    "ruff>=0.12.10",
]
